refactor: helper 메소드 변경
기존 JwtUtil 클래스는 helper 메소드임.
